Add backend lowering to Fx API.
