+++ projects/pt1/python/torch_mlir/compiler_utils.py
@@ -2,14 +2,59 @@
 # See https://llvm.org/LICENSE.txt for license information.
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 # Also available under a BSD-style license. See LICENSE.
-
+from enum import Enum
 from io import StringIO
 import os
 import sys
 import tempfile
+from typing import Union
 
 from torch_mlir.passmanager import PassManager
 from torch_mlir.ir import StringAttr
+
+
+class OutputType(Enum):
+    # This output type consists of `torch` dialect ops that have been converted
+    # maximally to value semantics, decomposed, and shapes have been inferred.
+    TORCH = "torch"
+
+    # The output type contains a mix of `linalg`-on-tensors ops, `scf`, and
+    # `arith` ops (and also `math` and `tm_tensor`). It can be thought of
+    # as taking the `TORCH` output type and lowering it so that tensor
+    # computations are done with `linalg`-on-tensors ops.
+    LINALG_ON_TENSORS = "linalg-on-tensors"
+
+    # This output type consists of `tosa` dialect ops. It can be thought of
+    # as taking the `TORCH` output type and lowering it to TOSA.
+    TOSA = "tosa"
+
+    # This output type consists of `stablehlo` dialect ops. It can be thought of
+    # as taking the `TORCH` output type and lowering it to StableHLO.
+    STABLEHLO = "stablehlo"
+
+    # Raw output of the JIT IR importer. This is not expected to be useful
+    # for end-users, but can be convenient for development or reporting bugs.
+    RAW = "raw"
+
+    @staticmethod
+    def get(spec: Union[str, "OutputType"]) -> "OutputType":
+        """Gets an OutputType from allowed way to specify one.
+
+        Args:
+          spec: An OutputType instance or the case-insensitive name of one of the
+            enum values.
+        Returns:
+          An OutputType instance.
+        """
+        if isinstance(spec, OutputType):
+            return spec
+        spec = spec.upper().replace("-", "_")
+        if spec not in OutputType.__members__:
+            raise ValueError(
+                f"For output_type= argument, expected one of: "
+                f"{', '.join(OutputType.__members__.keys())}"
+            )
+        return OutputType[spec]
 
 
 def get_module_name_for_debug_dump(module):
@@ -73,3 +118,50 @@ def run_pipeline_with_repro_report(module,
         raise TorchMlirCompilerError(trimmed_message) from None
     finally:
         sys.stderr = original_stderr
+
+
+def lower_mlir_module(verbose, output_type, module):
+    if verbose:
+        print("\n====================")
+        print("Torch Backend IR")
+        print(module)
+
+    if output_type == OutputType.TORCH:
+        return module
+
+    if output_type == OutputType.TOSA:
+        run_pipeline_with_repro_report(
+            module,
+            "builtin.module(torch-backend-to-tosa-backend-pipeline)",
+            "Lowering Torch Backend IR -> TOSA Backend IR",
+        )
+        if verbose:
+            print("\n====================")
+            print("TOSA Backend IR")
+            print(module)
+        return module
+
+    if output_type == OutputType.LINALG_ON_TENSORS:
+        run_pipeline_with_repro_report(
+            module,
+            "builtin.module(torch-backend-to-linalg-on-tensors-backend-pipeline)",
+            "Lowering Torch Backend IR -> Linalg-on-Tensors Backend IR",
+        )
+        if verbose:
+            print("\n====================")
+            print("LINALG Backend IR")
+            print(module)
+        return module
+
+    elif output_type == OutputType.STABLEHLO:
+        run_pipeline_with_repro_report(
+            module,
+            "builtin.module(torch-backend-to-stablehlo-backend-pipeline)",
+            "Lowering Torch Backend IR -> StableHLO Backend IR",
+        )
+        if verbose:
+            print("\n====================")
+            print("StableHLO Backend IR")
+            print(module)
+        return module
+    raise Exception(f"Unknown OutputType: {output_type}")
