+++ projects/pt1/python/torch_mlir/torchscript.py
@@ -4,7 +4,6 @@
 # Also available under a BSD-style license. See LICENSE.
 
 from typing import Optional, Sequence, Union, List, Dict, Tuple, Callable, Iterable
-from enum import Enum
 
 import sys
 from io import StringIO
@@ -17,63 +16,13 @@ import torch.fx
 from torch_mlir.dynamo import _get_decomposition_table
 from torch.fx.experimental.proxy_tensor import make_fx
 
-from .compiler_utils import run_pipeline_with_repro_report
+from .compiler_utils import (
+    OutputType,
+    run_pipeline_with_repro_report,
+    lower_mlir_module,
+)
 from torch_mlir.jit_ir_importer import ClassAnnotator, ImportOptions, ModuleBuilder
 from torch_mlir.jit_ir_importer.build_tools.library_generator import generate_library
-
-
-class OutputType(Enum):
-    """The kind of output that `torchscript.compile` can produce.
-
-    In MLIR terminology, this describes the mix of dialects that will be
-    produced by the conversion process.
-
-    In user-facing API's, this type can always be passed interchangeably with an
-    appropriate string specifying the output type. The allowed strings are
-    the set of enum vales, allowed to be case insensitive and with `-` allowed
-    in place of `_`. The `OutputType.get` static method can be used to convert
-    from a string to an `OutputType` instance.
-    """
-
-    # This output type consists of `torch` dialect ops that have been converted
-    # maximally to value semantics, decomposed, and shapes have been inferred.
-    TORCH = "torch"
-
-    # The output type contains a mix of `linalg`-on-tensors ops, `scf`, and
-    # `arith` ops (and also `math` and `tm_tensor`). It can be thought of
-    # as taking the `TORCH` output type and lowering it so that tensor
-    # computations are done with `linalg`-on-tensors ops.
-    LINALG_ON_TENSORS = "linalg-on-tensors"
-
-    # This output type consists of `tosa` dialect ops. It can be thought of
-    # as taking the `TORCH` output type and lowering it to TOSA.
-    TOSA = "tosa"
-
-    # This output type consists of `stablehlo` dialect ops. It can be thought of
-    # as taking the `TORCH` output type and lowering it to StableHLO.
-    STABLEHLO = "stablehlo"
-
-    # Raw output of the JIT IR importer. This is not expected to be useful
-    # for end-users, but can be convenient for development or reporting bugs.
-    RAW = "raw"
-
-    @staticmethod
-    def get(spec: Union[str, "OutputType"]) -> "OutputType":
-        """Gets an OutputType from allowed way to specify one.
-
-        Args:
-          spec: An OutputType instance or the case-insensitive name of one of the
-            enum values.
-        Returns:
-          An OutputType instance.
-        """
-        if isinstance(spec, OutputType):
-            return spec
-        spec = spec.upper().replace("-", "_")
-        if spec not in OutputType.__members__:
-            raise ValueError(f"For output_type= argument, expected one of: "
-                             f"{', '.join(OutputType.__members__.keys())}")
-        return OutputType[spec]
 
 
 class TensorPlaceholder:
@@ -270,56 +219,13 @@ def _canon_extra_library(extra_library, extra_library_file_name="custom_op_extra
         return ""
 
 
-def _lower_mlir_module(verbose, output_type, module):
-    if verbose:
-        print("\n====================")
-        print("Torch Backend IR")
-        print(module)
-
-    if output_type == OutputType.TORCH:
-        return module
-
-    if output_type == OutputType.TOSA:
-        run_pipeline_with_repro_report(
-            module, "builtin.module(torch-backend-to-tosa-backend-pipeline)",
-            "Lowering Torch Backend IR -> TOSA Backend IR")
-        if verbose:
-            print("\n====================")
-            print("TOSA Backend IR")
-            print(module)
-        return module
-
-    if output_type == OutputType.LINALG_ON_TENSORS:
-        run_pipeline_with_repro_report(
-            module,
-            "builtin.module(torch-backend-to-linalg-on-tensors-backend-pipeline)",
-            "Lowering Torch Backend IR -> Linalg-on-Tensors Backend IR")
-        if verbose:
-            print("\n====================")
-            print("LINALG Backend IR")
-            print(module)
-        return module
-
-    elif output_type == OutputType.STABLEHLO:
-        run_pipeline_with_repro_report(
-            module,
-            "builtin.module(torch-backend-to-stablehlo-backend-pipeline)",
-            "Lowering Torch Backend IR -> StableHLO Backend IR")
-        if verbose:
-            print("\n====================")
-            print("StableHLO Backend IR")
-            print(module)
-        return module
-    raise Exception(f"Unknown OutputType: {output_type}")
-
-
 def compile(model: torch.nn.Module,
             example_args: _example_args,
             output_type: Union[str, "OutputType"] = OutputType.TORCH,
             use_tracing: bool = False,
             ignore_traced_shapes=False,
             backend_legal_ops: Optional[Sequence[str]] = None,
-            extra_library: Iterable[Callable] = [],
+            extra_library: Optional[Iterable[Callable]] = None,
             verbose: bool = False,
             use_make_fx: bool = False,
             enable_ir_printing: bool = False):
@@ -363,6 +269,8 @@ def compile(model: torch.nn.Module,
         An MLIR module that contains the converted model in the specified
         output type.
     """
+    if extra_library is None:
+        extra_library = []
     extra_library_file_name = _canon_extra_library(extra_library)
     output_type = OutputType.get(output_type)
     example_args = ExampleArgs.get(example_args)
@@ -464,4 +372,4 @@ PyTorch TorchScript module -> torch-mlir Object Graph IR import failed with:
         enable_ir_printing=enable_ir_printing,
     )
 
-    return _lower_mlir_module(verbose, output_type, mb.module)
+    return lower_mlir_module(verbose, output_type, mb.module)
