+++ projects/pt1/python/torch_mlir_e2e_test/configs/fx_importer_backend.py
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 # Also available under a BSD-style license. See LICENSE.
 
-from typing import Union, Optional, Sequence
+from typing import Union
 
 import numpy as np
 import torch
@@ -12,14 +12,7 @@ from torch.export.graph_signature import OutputSpec, OutputKind
 from torch.export import ExportedProgram
 
 from torch_mlir import fx
-from torch_mlir.torchscript import (
-    _example_args,
-    OutputType,
-    BACKEND_LEGAL_OPS,
-    run_pipeline_with_repro_report,
-    _lower_mlir_module,
-    _canon_extra_library,
-)
+from torch_mlir.compiler_utils import OutputType
 from torch_mlir_e2e_test.configs.utils import (
     recursively_convert_to_numpy,
     recursively_convert_from_numpy,
@@ -39,44 +32,16 @@ def refine_result_type(_result):
 
 
 def jit(
-        prog: ExportedProgram,
-        func_name: str,
-        output_type: Union[str, "OutputType"] = OutputType.TORCH,
-        backend_legal_ops: Optional[Sequence[str]] = None,
-        extra_library=None,
-        verbose: bool = False,
+    prog: ExportedProgram,
+    func_name: str,
+    output_type: Union[str, OutputType] = OutputType.TORCH,
+    verbose: bool = False,
 ):
-    if extra_library is None:
-        extra_library = []
-    mlir_module = None
-
-    extra_library_file_name = _canon_extra_library(extra_library)
-    output_type = OutputType.get(output_type)
-    if backend_legal_ops is not None:
-        if output_type != OutputType.TORCH:
-            raise Exception("`backend_legal_ops` is only valid with the "
-                            "`torch` output type")
-        backend_legal_ops = list(sorted(set(backend_legal_ops)))
-    else:
-        backend_legal_ops = BACKEND_LEGAL_OPS.get(output_type, [])
-
-    option_string = ("{backend-legal-ops=" + ",".join(backend_legal_ops) +
-                     " extra-library=" + extra_library_file_name + "}")
-
-    mlir_module = fx.export_and_import(prog, func_name=func_name)
+    mlir_module = fx.export_and_import(
+        prog, output_type=output_type, func_name=func_name, verbose=verbose
+    )
     assert mlir_module is not None
-    run_pipeline_with_repro_report(
-        mlir_module,
-        f"builtin.module(torch-simplification-pipeline)",
-        "Simplification pipeline for torch dialect",
-    )
-    run_pipeline_with_repro_report(
-        mlir_module,
-        f"builtin.module(torch-function-to-torch-backend-pipeline{option_string})",
-        "Lowering TorchFX IR -> Torch Backend IR",
-    )
-
-    return _lower_mlir_module(verbose, output_type, mlir_module)
+    return mlir_module
 
 
 class FxImporterTestConfig(TestConfig):
